+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/views_infra_services.cljs
@@ -8,8 +8,7 @@
     [sixsq.nuvla.ui.utils.form-fields :as ff]
     [sixsq.nuvla.ui.utils.semantic-ui :as ui]
     [sixsq.nuvla.ui.utils.semantic-ui-extensions :as uix]
-    [sixsq.nuvla.ui.utils.style :as style]
-    [sixsq.nuvla.ui.utils.time :as time]))
+    [sixsq.nuvla.ui.utils.style :as style]))
 
 (defn summary-row
   []