+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/views_summary.cljs
@@ -8,11 +8,11 @@
     [sixsq.nuvla.ui.deployment-dialog.views-files :as files-step]
     [sixsq.nuvla.ui.deployment-dialog.views-infra-services :as infra-services-step]
     [sixsq.nuvla.ui.deployment-dialog.views-license :as license-step]
+    [sixsq.nuvla.ui.deployment-dialog.views-module-version :as module-version-step]
     [sixsq.nuvla.ui.deployment-dialog.views-price :as price-step]
     [sixsq.nuvla.ui.deployment-dialog.views-registries :as registries-step]
     [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
-    [sixsq.nuvla.ui.utils.semantic-ui :as ui]
-    [sixsq.nuvla.ui.deployment-dialog.views-module-version :as module-version-step]))
+    [sixsq.nuvla.ui.utils.semantic-ui :as ui]))
 
 
 (defn application-row