+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Updated
 
-  :todo - Deployment - Remove fetch button from deployment detail page
+  - Deployment detail - Remove fetch button
   - Deployment modal - Add version selection section
 
 ## [2.10.3] - 2020-11-17